backend/pythonBackend/requirements.txt
# FastAPI Backend with Strands AI Agent Integration

# Core web framework
fastapi==0.118.3
uvicorn==0.37.0
starlette==0.48.0

# Database
motor==3.7.1
pymongo==4.15.3

# Environment and configuration
python-dotenv==1.1.1
pydantic==2.12.0
pydantic-settings==2.11.0

# AWS and Strands AI
boto3==1.40.50
botocore==1.40.50
strands-agents==1.12.0
strands-agents-builder==0.1.10
strands-agents-tools==0.2.11

# HTTP and networking
httpx==0.28.1
httpx-sse==0.4.3
httpcore==1.0.9
requests==2.32.5

# Image processing
Pillow==11.2.1

# Data processing and validation
jsonschema==4.25.1
jsonschema-specifications==2025.9.1

# Text processing
beautifulsoup4==4.14.2
markdown-it-py==4.0.0
markdownify==1.2.0

# CLI and terminal
click==8.3.0
rich==14.2.0
prompt_toolkit==3.0.52

# Utilities
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
h11==0.16.0
idna==3.10
python-dateutil==2.9.0.post0
python-multipart==0.0.20
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0

# Optional: Slack integration
slack_bolt==1.26.0
slack_sdk==3.37.0


# Optional: Mathematical computations
sympy==1.14.0

# Optional: MCP (Model Context Protocol)
mcp==1.17.0

# Optional: Server-sent events
sse-starlette==3.0.2

# Optional: JWT tokens
PyJWT==2.10.1

# Optional: OpenTelemetry monitoring
opentelemetry-api==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0